requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
numpy
pillow
ultralytics
albumentations
python-multipart
